moved onoff sensor to binary_sensor
