+++ homeassistant/components/enocean/binary_sensor.py
@@ -2,11 +2,22 @@
 import voluptuous as vol
 
 from homeassistant.components.binary_sensor import (
+    DEVICE_CLASS_MOTION,
+    DEVICE_CLASS_WINDOW,
     DEVICE_CLASSES_SCHEMA,
     PLATFORM_SCHEMA,
     BinarySensorEntity,
 )
-from homeassistant.const import CONF_DEVICE_CLASS, CONF_ID, CONF_NAME
+from homeassistant.const import (
+    CONF_DEVICE_CLASS,
+    CONF_ID,
+    CONF_NAME,
+    DEVICE_CLASS_BATTERY,
+    STATE_CLOSED,
+    STATE_OFF,
+    STATE_ON,
+    STATE_OPEN,
+)
 import homeassistant.helpers.config_validation as cv
 
 from .device import EnOceanEntity
@@ -15,11 +26,42 @@ DEFAULT_NAME = "EnOcean binary sensor"
 DEPENDENCIES = ["enocean"]
 EVENT_BUTTON_PRESSED = "button_pressed"
 
+CONF_INVERTED = "inverted"
+
+SENSOR_TYPE_BATTERY = "battery"
+SENSOR_TYPE_BUTTON_PRESSED = "button"
+SENSOR_TYPE_MOTION = "motion"
+SENSOR_TYPE_WINDOW = "window"
+
+SENSOR_TYPES = {
+    SENSOR_TYPE_BATTERY: {
+        "name": "Battery state",
+        "icon": "mdi:battery",
+        "class": DEVICE_CLASS_BATTERY,
+    },
+    SENSOR_TYPE_BUTTON_PRESSED: {
+        "name": "Button pressed",
+        "icon": "mdi:gesture-tap-button",
+        "class": None,
+    },
+    SENSOR_TYPE_MOTION: {
+        "name": "Motion",
+        "icon": "mdi:motion",
+        "class": DEVICE_CLASS_MOTION,
+    },
+    SENSOR_TYPE_WINDOW: {
+        "name": "Window",
+        "icon": "mdi:window",
+        "class": DEVICE_CLASS_WINDOW,
+    },
+}
+
 PLATFORM_SCHEMA = PLATFORM_SCHEMA.extend(
     {
         vol.Required(CONF_ID): vol.All(cv.ensure_list, [vol.Coerce(int)]),
         vol.Optional(CONF_NAME, default=DEFAULT_NAME): cv.string,
         vol.Optional(CONF_DEVICE_CLASS): DEVICE_CLASSES_SCHEMA,
+        vol.Optional(CONF_INVERTED, default=0): cv.boolean,
     }
 )
 
@@ -28,9 +70,25 @@ def setup_platform(hass, config, add_entities, discovery_info=None):
     """Set up the Binary Sensor platform for EnOcean."""
     dev_id = config.get(CONF_ID)
     dev_name = config.get(CONF_NAME)
-    device_class = config.get(CONF_DEVICE_CLASS)
 
-    add_entities([EnOceanBinarySensor(dev_id, dev_name, device_class)])
+    sensor_type = config.get(CONF_DEVICE_CLASS)
+
+    if (
+        sensor_type == SENSOR_TYPE_BATTERY
+        or sensor_type == SENSOR_TYPE_BUTTON_PRESSED
+        or sensor_type == SENSOR_TYPE_MOTION
+    ):
+        inverted = config.get(CONF_INVERTED)
+        add_entities(
+            [EnOceanOnOffSensor(dev_id, dev_name, sensor_type, inverted=inverted)]
+        )
+    elif sensor_type == SENSOR_TYPE_WINDOW:
+        inverted = config.get(CONF_INVERTED)
+        add_entities(
+            [EnOceanOpenClosedSensor(dev_id, dev_name, sensor_type, inverted=inverted)]
+        )
+    else:
+        add_entities([EnOceanBinarySensor(dev_id, dev_name, sensor_type)])
 
 
 class EnOceanBinarySensor(EnOceanEntity, BinarySensorEntity):
@@ -107,4 +165,102 @@ class EnOceanBinarySensor(EnOceanEntity, BinarySensorEntity):
                 "which": self.which,
                 "onoff": self.onoff,
             },
+        )
+
+
+class EnOceanOnOffSensor(EnOceanBinarySensor):
+    """Representation of an EnOcean on-off sensor device, storing state in data byte 0.0, most often part of a multi-sensor device.
+
+    EEPs (EnOcean Equipment Profiles):
+    - D5-00-01
+    - A5-10-02 (slide switch of operating panel)
+    - A5-10-06 (slide switch of operating panel)
+    - A5-10-09 (slide switch of operating panel)
+    - A5-10-0D (slide switch of operating panel)
+    - A5-10-11 (slide switch of operating panel)
+    - A5-10-14 (slide switch of operating panel)
+    - A5-10-20 (user intervention on device)
+    - A5-10-21 (user intervention on device)
+    - A5-10-21 (occupied)
+    - A5-11-02 (occupied)
+    - A5-11-04 (light on)
+    - A5-14-08 (vibration detected)
+    - A5-14-0A (vibration detected)
+    - A5-20-10 (HVAC unit running state)
+    - A5-20-11 (HVAC alarm error state)
+    - A5-38-08 (switching command)
+
+    For the following EEPs the inverted flag has to be set to true:
+    - A5-08-01 (occupancy button pressed)
+    - A5-08-03 (occupancy button pressed)
+    - A5-10-01 (occupancy button pressed)
+    - A5-10-05 (occupancy button pressed)
+    - A5-10-08 (occupancy button pressed)
+    - A5-10-0C (occupancy button pressed)
+    - A5-10-0D (occupancy button pressed)
+    - A5-10-10 (occupancy button pressed)
+    - A5-10-13 (occupancy button pressed)
+    - A5-10-16 (occupancy button pressed)
+    - A5-10-17 (occupancy button pressed)
+    - A5-10-18 (occupancy button pressed)
+    - A5-10-19 (room occupancied)
+    - A5-10-1A (occupancy button pressed)
+    - A5-10-1B (occupancy button pressed)
+    - A5-10-1C (occupancy button pressed)
+    - A5-10-1D (occupancy button pressed)
+    - A5-10-1F (occupancy button pressed)
+    - A5-13-07 (battery low=0, battery ok=1)
+    - A5-13-08 (battery low=0, battery ok=1)
+    - A5-20-12 (room occupancied)
+    """
+
+    def __init__(
+        self,
+        dev_id,
+        dev_name,
+        sensor_type,
+        state_on=STATE_ON,
+        state_off=STATE_OFF,
+        inverted: bool = False,
+    ):
+        """Initialize the EnOcean on-off sensor device."""
+        super().__init__(dev_id, dev_name, sensor_type)
+        self._state_on = state_on
+        self._state_off = state_off
+        self._inverted = inverted
+
+    def value_changed(self, packet):
+        """Update the internal state of the sensor."""
+        stateOn = (packet.data[0] & 0x01) == 0x01
+
+        if stateOn and not self._inverted:
+            self._state = self._state_on
+        else:
+            self._state = self._state_off
+
+        self.schedule_update_ha_state()
+
+
+class EnOceanOpenClosedSensor(EnOceanOnOffSensor):
+    """Represents an EnOcean Open-Closed sensor device.
+
+    EEPs (EnOcean Equipment Profiles):
+    - D5-00-01
+
+    For the following EEPs the inverted flag has to be set to true:
+    - A5-10-0A (contact state, 1=Open)
+    - A5-10-08 (contact state, 1=Open)
+    - A5-14-01 to A5-14-04 (contact state, 1=Open)
+    - A5-30-02 (contact state, 1=Open)
+    """
+
+    def __init__(self, dev_id, dev_name, sensor_type, inverted: bool = False):
+        """Initialize EnOceanOpenClosedSensor."""
+        super().__init__(
+            dev_id,
+            dev_name,
+            sensor_type,
+            state_on=STATE_OPEN,
+            state_off=STATE_CLOSED,
+            inverted=inverted,
         )
